TransformImage legacy code cleanup
