+++ app/src/main/java/fr/free/nrw/commons/edit/EditViewModel.kt
@@ -2,66 +2,102 @@ package fr.free.nrw.commons.edit
 
 import android.content.Context
 import android.net.Uri
+import androidx.core.net.toUri
 import androidx.lifecycle.ViewModel
+import fr.free.nrw.commons.ajpegtran.Jpegtran
 import fr.free.nrw.commons.ajpegtran.Properties
 import fr.free.nrw.commons.ajpegtran.blur.BlurRegion
+import fr.free.nrw.commons.ajpegtran.rotate.RotationDegree
+import timber.log.Timber
 import java.io.File
 
 /**
- * ViewModel for image editing operations.
+ * ViewModel for image editing operations using Jpegtran.
  */
 class EditViewModel : ViewModel() {
-    // Ideally should be injected using DI
-    private val transformImage: TransformImage = TransformImageImpl()
+    private var jpegtran: Jpegtran? = null
 
     /**
      * Initialize the single Jpegtran instance for this editing session.
      * Must be called with Application Context from EditActivity.
      */
     fun initJpegtran(context: Context, imagePath: String) {
-        transformImage.initJpegtran(context, imagePath)
+        if (jpegtran == null) {
+            val imageUri = if (imagePath.startsWith("content://") || imagePath.startsWith("file://")) {
+                imagePath.toUri()
+            } else {
+                File(imagePath).toUri()
+            }
+            jpegtran = Jpegtran(context, imageUri)
+        }
     }
 
     /**
      * Returns properties of the JPEG image.
      */
     fun getProperties(uri: Uri): Properties {
-        return transformImage.getProperties(uri)
+        return jpegtran?.getProperties(uri)
+            ?: throw IllegalStateException("Jpegtran not initialized")
     }
 
     /**
      * Clear and cleanup temporary files in Jpegtran.
      */
+    fun cleanup() {
+        jpegtran?.cleanup()
+        jpegtran = null
+    }
+
     override fun onCleared() {
-        // If user canceled/backed or finishes editing, clear the temporary files in the library.
-        transformImage.cleanup()
+        cleanup()
         super.onCleared()
     }
 
     /**
      * Rotates the specified image file by the given degree.
      *
-     * @param degree The degree by which to rotate the image.
      * @param imageFile The File representing the image to be rotated.
+     * @param degree The degree by which to rotate the image.
      * @param savePath The directory to save the rotated image in.
-     * @return The rotated image File, or null if the rotation operation fails.
+     * @return The rotated image File.
      */
     fun rotateImage(
         imageFile: File,
         degree: Int,
         savePath: File
     ): File {
-        return transformImage.rotateImage(imageFile, degree, savePath)
+        Timber.tag("Trying to rotate image").d("Starting")
+        val imagePath = "cropped_${System.currentTimeMillis()}.jpg"
+        val output = File(savePath, imagePath)
+        val normalizedDegree = ((degree % 360) + 360) % 360
+        if (normalizedDegree == 0) {
+            imageFile.copyTo(output, overwrite = true)
+            return output
+        }
+        val rotationDegree = when (normalizedDegree) {
+            90 -> RotationDegree.ROTATE_90
+            180 -> RotationDegree.ROTATE_180
+            270 -> RotationDegree.ROTATE_270
+            else -> throw IllegalArgumentException("Unsupported degree: $degree")
+        }
+        try {
+            jpegtran!!.rotate(rotationDegree)
+            jpegtran!!.save(output.toUri())
+            return output
+        } catch (e: Exception) {
+            Timber.e(e, "saveEditedImage: Failed to rotate image")
+            throw e
+        }
     }
 
     /**
-     * Crops the specified image file using lossless JPEG cropping.
+     * Crops the specified image file using lossless JPEG cropping via Jpegtran.
      *
      * @param left The left coordinate of the crop rectangle.
      * @param top The top coordinate of the crop rectangle.
      * @param width The width of the crop rectangle.
      * @param height The height of the crop rectangle.
-     * @return The cropped image File, or null if the crop operation fails.
+     * @return The cropped image File.
      */
     fun cropImage(
         left: Int,
@@ -69,7 +105,26 @@ class EditViewModel : ViewModel() {
         width: Int,
         height: Int,
         savePath: File,
-    ): File = transformImage.cropImage(left, top, width, height, savePath)
+    ): File {
+        Timber.tag("Trying to crop image").d(
+            "Starting crop: left=$left, top=$top, width=$width, height=$height"
+        )
+        val imagePath = "cropped_${System.currentTimeMillis()}.jpg"
+        val output = File(savePath, imagePath)
+        try {
+            jpegtran!!.crop(
+                width,
+                height,
+                left,
+                top
+            )
+            jpegtran!!.save(output.toUri())
+            return output
+        } catch (e: Exception) {
+            Timber.e(e, "saveEditedImage: Failed to crop image")
+            throw e
+        }
+    }
 
     /**
      * Blurs the specified regions of the image file.
@@ -81,5 +136,17 @@ class EditViewModel : ViewModel() {
     fun blurImage(
         regions: List<BlurRegion>,
         savePath: File
-    ): File = transformImage.blurImage(regions, savePath)
+    ): File {
+        Timber.tag("Trying to blur image").d("Starting blur of ${regions.size} regions")
+        val imagePath = "blurred_${System.currentTimeMillis()}.jpg"
+        val output = File(savePath, imagePath)
+        try {
+            jpegtran!!.blur(regions)
+            jpegtran!!.save(output.toUri())
+        } catch (e: Exception) {
+            Timber.e(e, "blurImage: Failed to blur image")
+            throw e
+        }
+        return output
+    }
 }
